+++ apps/ledger-live-desktop/src/renderer/families/cosmos/DelegationFlowModal/Info/index.jsx
@@ -2,15 +2,13 @@
 import React, { useCallback } from "react";
 import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
-
 import type { Account, AccountLike } from "@ledgerhq/types-live";
-
 import { openModal, closeModal } from "~/renderer/actions/modals";
 import EarnRewardsInfoModal from "~/renderer/components/EarnRewardsInfoModal";
 import WarnBox from "~/renderer/components/WarnBox";
-import { urls } from "~/config/urls";
 import { openURL } from "~/renderer/linking";
 import LinkWithExternalIcon from "~/renderer/components/LinkWithExternalIcon";
+import cryptoFactory from "@ledgerhq/live-common/families/cosmos/chain/chain";
 
 type Props = {
   name?: string,
@@ -31,24 +29,33 @@ export default function CosmosEarnRewardsInfoModal({ name, account, parentAccoun
     );
   }, [parentAccount, account, dispatch, name]);
 
-  const onLearnMore = useCallback(() => {
-    openURL(urls.cosmosStakingRewards);
+  const onLearnMore = useCallback((currencyId: string) => {
+    openURL(cryptoFactory(currencyId).stakingDocUrl);
   }, []);
-
+  const crypto = cryptoFactory(account.currency.id);
   return (
     <EarnRewardsInfoModal
       name={name}
       onNext={onNext}
-      description={t("cosmos.delegation.flow.steps.starter.description")}
+      description={t("cosmos.delegation.flow.steps.starter.description", {
+        currencyTicker: account.currency.ticker,
+      })}
       bullets={[
         t("cosmos.delegation.flow.steps.starter.bullet.0"),
         t("cosmos.delegation.flow.steps.starter.bullet.1"),
-        t("cosmos.delegation.flow.steps.starter.bullet.2"),
+        t("cosmos.delegation.flow.steps.starter.bullet.2", {
+          numberOfDays: crypto.unbondingPeriod,
+        }),
       ]}
       additional={
         <WarnBox>{t("cosmos.delegation.flow.steps.starter.warning.description")}</WarnBox>
       }
-      footerLeft={<LinkWithExternalIcon label={t("delegation.howItWorks")} onClick={onLearnMore} />}
+      footerLeft={
+        <LinkWithExternalIcon
+          label={t("delegation.howItWorks")}
+          onClick={() => onLearnMore(account.currency.id)}
+        />
+      }
     />
   );
 }