+++ apps/ledger-live-desktop/src/renderer/families/tron/AccountHeaderManageActions.js
@@ -1,5 +1,9 @@
 // @flow
-import { getAccountUnit, getMainAccount } from "@ledgerhq/live-common/account/index";
+import {
+  getAccountUnit,
+  getMainAccount,
+  isAccountEmpty,
+} from "@ledgerhq/live-common/account/index";
 import type { Account, AccountLike } from "@ledgerhq/types-live";
 import { useCallback } from "react";
 import { useTranslation } from "react-i18next";
@@ -18,6 +22,7 @@ type Props = {
 const AccountHeaderManageActionsComponent = ({ account, parentAccount }: Props) => {
   const { t } = useTranslation();
   const dispatch = useDispatch();
+
   const locale = useSelector(localeSelector);
   const unit = getAccountUnit(account);
   const mainAccount = getMainAccount(account, parentAccount);
@@ -28,7 +33,14 @@ const AccountHeaderManageActionsComponent = ({ account, parentAccount }: Props) 
   const earnRewardDisabled = tronPower === 0 && spendableBalance.lt(minAmount);
 
   const onClick = useCallback(() => {
-    if (tronPower > 0) {
+    if (isAccountEmpty(account) || earnRewardDisabled) {
+      dispatch(
+        openModal("MODAL_NO_FUNDS_STAKE", {
+          account,
+          parentAccount,
+        }),
+      );
+    } else if (tronPower > 0) {
       dispatch(
         openModal("MODAL_MANAGE_TRON", {
           parentAccount,
@@ -43,7 +55,7 @@ const AccountHeaderManageActionsComponent = ({ account, parentAccount }: Props) 
         }),
       );
     }
-  }, [dispatch, tronPower, account, parentAccount]);
+  }, [account, earnRewardDisabled, tronPower, dispatch, parentAccount]);
 
   if (parentAccount) return null;
 
@@ -62,7 +74,6 @@ const AccountHeaderManageActionsComponent = ({ account, parentAccount }: Props) 
     {
       key: "Stake",
       onClick: onClick,
-      disabled: earnRewardDisabled,
       icon: IconCoins,
       label: t("account.stake"),
       tooltip: disabledLabel,