+++ apps/ledger-live-mobile/src/families/cosmos/operationDetails.tsx
@@ -12,17 +12,25 @@ import { BigNumber } from "bignumber.js";
 import { getAccountUnit } from "@ledgerhq/live-common/account/helpers";
 import type { CosmosDelegationInfo } from "@ledgerhq/live-common/families/cosmos/types";
 import { useSelector } from "react-redux";
+import cryptoFactory from "@ledgerhq/live-common/families/cosmos/chain/chain";
 import Section from "../../screens/OperationDetails/Section";
-import { urls } from "../../config/urls";
 import { discreetModeSelector, localeSelector } from "../../reducers/settings";
 
-function getURLFeesInfo(op: Operation): string | null | undefined {
-  return op.fee.gt(200000) ? urls.cosmosStakingRewards : undefined;
+function getURLFeesInfo(
+  op: Operation,
+  currencyId: string,
+): string | null | undefined {
+  return op.fee.gt(200000)
+    ? cryptoFactory(currencyId).stakingDocUrl
+    : undefined;
 }
 
-function getURLWhatIsThis(op: Operation): string | null | undefined {
+function getURLWhatIsThis(
+  op: Operation,
+  currencyId: string,
+): string | null | undefined {
   return op.type !== "IN" && op.type !== "OUT"
-    ? urls.cosmosStakingRewards
+    ? cryptoFactory(currencyId).stakingDocUrl
     : undefined;
 }