+++ apps/ledger-live-mobile/src/families/elrond/components/Body/components/Delegations/components/Right/types.ts
@@ -1,0 +1,4 @@
+export interface RightPropsType {
+  disabled?: boolean;
+  onPress: () => void;
+}