+++ apps/ledger-live-mobile/src/families/elrond/components/Body/components/Rewards/types.ts
@@ -1,0 +1,7 @@
+import type { ElrondAccount } from "@ledgerhq/live-common/families/elrond/types";
+import type { DelegationType } from "../../../../types";
+
+export interface RewardsPropsType {
+  account: ElrondAccount;
+  delegations: DelegationType[];
+}