+++ apps/ledger-live-mobile/src/families/elrond/components/Flows/Delegate/components/PickValidator/components/Item/styles.ts
@@ -1,0 +1,37 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  validator: {
+    flexDirection: "row",
+    alignItems: "center",
+    height: 56,
+  },
+  validatorBody: {
+    flex: 1,
+    flexDirection: "column",
+    marginLeft: 12,
+  },
+  validatorName: {
+    fontSize: 14,
+  },
+  overdelegatedIndicator: {
+    position: "absolute",
+    width: 10,
+    height: 10,
+    borderRadius: 10,
+    top: 34,
+    left: 24,
+    borderWidth: 1,
+  },
+  overdelegated: {
+    fontSize: 12,
+  },
+  validatorYield: {
+    fontSize: 14,
+  },
+  validatorYieldFull: {
+    opacity: 0.5,
+  },
+});
+
+export default styles;