+++ apps/ledger-live-mobile/src/reducers/settings.ts
@@ -158,6 +158,7 @@ export const INITIAL_STATE: SettingsState = {
     areNotificationsAllowed: true,
     announcementsCategory: true,
     recommendationsCategory: true,
+    largeMoverCategory: true,
   },
   walletTabNavigatorLastVisitedTab: ScreenName.Portfolio,
   displayStatusCenter: false,
@@ -591,6 +592,7 @@ const handlers: ReducerMap<SettingsState, SettingsPayload> = {
     displayStatusCenter: (action as Action<SettingsSetStatusCenterPayload>)
       .payload.displayStatusCenter,
   }),
+
   [SettingsActionTypes.SET_OVERRIDDEN_FEATURE_FLAG]: (state, action) => {
     const {
       payload: { id, value },