+++ libs/ledger-live-common/src/families/algorand/account.ts
@@ -2,7 +2,8 @@ import invariant from "invariant";
 import { BigNumber } from "bignumber.js";
 import { getAccountUnit } from "../../account";
 import { formatCurrencyUnit } from "../../currencies";
-import type { AlgorandAccount, AlgorandResources } from "./types";
+import type { Account } from "@ledgerhq/types-live";
+import type { AlgorandResources, AlgorandAccount } from "./types";
 import type { Operation } from "@ledgerhq/types-live";
 import type { Unit } from "@ledgerhq/types-cryptoassets";
 
@@ -24,8 +25,8 @@ function formatOperationSpecifics(
     : "";
 }
 
-function formatAccountSpecifics(account: AlgorandAccount): string {
-  const { algorandResources } = account;
+function formatAccountSpecifics(account: Account): string {
+  const { algorandResources } = account as AlgorandAccount;
   invariant(algorandResources, "algorand account expected");
   const unit = getAccountUnit(account);
   const formatConfig = {