+++ libs/ledger-live-common/src/network.ts
@@ -1,54 +1,15 @@
 import invariant from "invariant";
-import axios, { AxiosResponse } from "axios";
+import axios from "axios";
 import type { AxiosError, AxiosRequestConfig } from "axios";
 import { log } from "@ledgerhq/logs";
 import { NetworkDown, LedgerAPI5xx, LedgerAPI4xx } from "@ledgerhq/errors";
 import { retry } from "./promise";
-import { getEnv } from "./env";
+import { getEnv, changes } from "./env";
 
-type Metadata = { startTime: number };
-type ExtendedXHRConfig = AxiosRequestConfig & { metadata?: Metadata };
-
-export const requestInterceptor = (
-  request: AxiosRequestConfig
-): ExtendedXHRConfig => {
-  const { baseURL, url, method = "", data } = request;
-  log("network", `${method} ${baseURL || ""}${url}`, { data });
-
-  // $FlowFixMe (LLD side)
-  const req: ExtendedXHRConfig = request;
-
-  req.metadata = {
-    startTime: Date.now(),
-  };
-
-  return req;
-};
-
-export const responseInterceptor = (
-  response: {
-    config: ExtendedXHRConfig;
-  } & AxiosResponse<any>
-) => {
-  const { baseURL, url, method = "", metadata } = response.config;
-  const { startTime = 0 } = metadata || {};
-
-  log(
-    "network-success",
-    `${response.status} ${method} ${baseURL || ""}${url} (${(
-      Date.now() - startTime
-    ).toFixed(0)}ms)`,
-    getEnv("DEBUG_HTTP_RESPONSE") ? { data: response.data } : undefined
-  );
-
-  return response;
-};
-
-export const errorInterceptor = (error: AxiosError<any>) => {
-  const config = error?.response?.config as ExtendedXHRConfig | null;
+export const errorInterceptor = (error: AxiosError<any>): AxiosError<any> => {
+  const config = error?.response?.config || null;
   if (!config) throw error;
-  const { baseURL, url, method = "", metadata } = config;
-  const { startTime = 0 } = metadata || {};
+  const { baseURL, url, method = "" } = config;
 
   let errorToThrow;
   if (error.response) {
@@ -73,28 +34,18 @@ export const errorInterceptor = (error: AxiosError<any>) => {
     }
     log(
       "network-error",
-      `${status} ${method} ${baseURL || ""}${url} (${(
-        Date.now() - startTime
-      ).toFixed(0)}ms): ${errorToThrow.message}`,
+      `${status} ${method} ${baseURL || ""}${url}: ${errorToThrow.message}`,
       getEnv("DEBUG_HTTP_RESPONSE") ? { data: data } : {}
     );
     throw errorToThrow;
   } else if (error.request) {
-    log(
-      "network-down",
-      `DOWN ${method} ${baseURL || ""}${url} (${(
-        Date.now() - startTime
-      ).toFixed(0)}ms)`
-    );
+    log("network-down", `DOWN ${method} ${baseURL || ""}${url}`);
     throw new NetworkDown();
   }
   throw error;
 };
 
-axios.interceptors.request.use(requestInterceptor);
-
-// $FlowFixMe LLD raise issues here
-axios.interceptors.response.use(responseInterceptor, errorInterceptor);
+axios.interceptors.response.use(undefined, errorInterceptor);
 
 const makeError = (msg, status, url, method) => {
   const obj = {
@@ -160,5 +111,20 @@ const implementation = (arg: AxiosRequestConfig): Promise<any> => {
 
   return promise;
 };
+
+// attach the env "LEDGER_CLIENT_VERSION" to set the header globally for axios
+function setAxiosLedgerClientVersionHeader(value: string) {
+  if (value) {
+    axios.defaults.headers.common["X-Ledger-Client-Version"] = value;
+  } else {
+    delete axios.defaults.headers.common["X-Ledger-Client-Version"];
+  }
+}
+setAxiosLedgerClientVersionHeader(getEnv("LEDGER_CLIENT_VERSION"));
+changes.subscribe((e) => {
+  if (e.name === "LEDGER_CLIENT_VERSION") {
+    setAxiosLedgerClientVersionHeader(e.value);
+  }
+});
 
 export default implementation;